Moved `SyncSendPointer` and `SendPointer` to 'shadow-shim-helper-rs'
