+++ src/main/core/work/task.rs
@@ -60,13 +60,11 @@ impl PartialEq for TaskRef {
 impl Eq for TaskRef {}
 
 pub mod export {
+    use shadow_shim_helper_rs::util::SyncSendPointer;
     use shadow_shim_helper_rs::{notnull::notnull_mut, HostId};
 
     use super::*;
-    use crate::{
-        host::host::Host,
-        utility::{HostTreePointer, SyncSendPointer},
-    };
+    use crate::{host::host::Host, utility::HostTreePointer};
 
     pub type TaskCallbackFunc = extern "C" fn(*const Host, *mut libc::c_void, *mut libc::c_void);
     pub type TaskObjectFreeFunc = Option<extern "C" fn(*mut libc::c_void)>;
