+++ src/main/utility/mod.rs
@@ -34,69 +34,6 @@ use shadow_shim_helper_rs::HostId;
 use crate::core::worker::Worker;
 use crate::host::host::Host;
 
-/// A type that allows us to make a pointer Send + Sync since there is no way
-/// to add these traits to the pointer itself.
-#[derive(Debug)]
-pub struct SyncSendPointer<T>(*mut T);
-
-// We can't automatically `derive` Copy and Clone without unnecessarily
-// requiring T to be Copy and Clone.
-// https://github.com/rust-lang/rust/issues/26925
-impl<T> Copy for SyncSendPointer<T> {}
-impl<T> Clone for SyncSendPointer<T> {
-    fn clone(&self) -> Self {
-        Self(self.0)
-    }
-}
-
-unsafe impl<T> Send for SyncSendPointer<T> {}
-unsafe impl<T> Sync for SyncSendPointer<T> {}
-
-impl<T> SyncSendPointer<T> {
-    /// # Safety
-    ///
-    /// The object pointed to by `ptr` must actually be `Sync` and `Send` or
-    /// else not subsequently used in contexts where it matters.
-    pub unsafe fn new(ptr: *mut T) -> Self {
-        Self(ptr)
-    }
-
-    pub fn ptr(&self) -> *mut T {
-        self.0
-    }
-}
-
-/// A type that allows us to make a pointer Send since there is no way
-/// to add this traits to the pointer itself.
-#[derive(Debug)]
-pub struct SendPointer<T>(*mut T);
-
-// We can't automatically `derive` Copy and Clone without unnecessarily
-// requiring T to be Copy and Clone.
-// https://github.com/rust-lang/rust/issues/26925
-impl<T> Copy for SendPointer<T> {}
-impl<T> Clone for SendPointer<T> {
-    fn clone(&self) -> Self {
-        Self(self.0)
-    }
-}
-
-unsafe impl<T> Send for SendPointer<T> {}
-
-impl<T> SendPointer<T> {
-    /// # Safety
-    ///
-    /// The object pointed to by `ptr` must actually be `Send` or else not
-    /// subsequently used in contexts where it matters.
-    pub unsafe fn new(ptr: *mut T) -> Self {
-        Self(ptr)
-    }
-
-    pub fn ptr(&self) -> *mut T {
-        self.0
-    }
-}
-
 /// A pointer to an object that is safe to dereference from any thread,
 /// *if* the Host lock for the specified host is held.
 #[derive(Debug)]
